- Added a backup system (written in Python) for the Trac Wiki pages (excluding the default ones, in other words: only the project-related content).
- Added the Wiki pages in their current state.

This is an additional measure to the backups made daily by CVS-Dude.com
Because it is a manual procedure, it won't record all the changes in the Wiki, but at least offers some kind of redundancy, and it's free ! :) (unlike the backup-download service of CVS-Dude).

From now-on, we can launch the Python script regularly (as soon as there are significant changes in the Wiki), in order to let SVN register the changes.