- optimization: event args cache for PropertyChanged notifications + fast cache (similar proprietary lightweight implementation) for Property dependencies - started to open-up the base class that handles PropertyChanged notifications, in order to make it reusable not only as a base class (which forces sub-classing)
This refactoring currently works fine with the popup modal window implementation. Next step is to check it with the AudioPane, as it's where there are the most commands and properties at the moment (oh yeah, another improvement to come is the suppression of AutomaticRequerySuggested in favor of a built-in and more lightweight event mechanism).